requirements.txt
python-dotenv
requests
pyTelegramBotAPI

